requirements.txt
alembic==1.14.0
attrs==24.3.0
blinker==1.9.0
click==8.1.8
coverage==5.5
ecdsa==0.19.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.6.1
pytest==6.2.5
python-dotenv==0.19.0
python-jose==3.3.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
toml==0.10.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
